integration-tests: rename to policy_test
